refactor: relocate access functions
