+++ src/access/index.ts
@@ -1,0 +1,3 @@
+export * from './isAdmin';
+export * from './isAdminOrEditor';
+export * from './isAdminOrSelf';
